Split TASTy reflect interface and implementation
Mostly to simplify the maintainability and documentation of distinct parts of the implementation.
Abstract types are defined separately in TastyCore to avoid cake pattern (except for Printers).
Extracted the `show` extension methods to avoid cakes.
