+++ library/src/scala/tasty/reflect/ContextOps.scala
@@ -1,0 +1,19 @@
+package scala.tasty
+package reflect
+
+trait ContextOps extends TastyCore {
+
+  trait ContextAPI {
+    def owner: Definition
+
+    /** Returns the source file being compiled. The path is relative to the current working directory. */
+    def source: java.nio.file.Path
+  }
+  implicit def ContextDeco(ctx: Context): ContextAPI
+
+  implicit def rootContext: Context
+
+  /** Root position of this tasty context. For macros it corresponds to the expansion site. */
+  def rootPosition: Position
+
+}
