+++ library/src/scala/tasty/reflect/TastyCore.scala
@@ -1,0 +1,151 @@
+package scala.tasty.reflect
+
+/** Tasty reflect abstract types
+ *
+ *  ```none
+ *
+ *  +- Tree -+- PackageClause
+ *           +- Import
+ *           +- Statement -+- Definition --+- ClassDef
+ *                         |               +- TypeDef
+ *                         |               +- DefDef
+ *                         |               +- ValDef
+ *                         |               +- PackageDef
+ *                         |
+ *                         +- Term --------+- Ident
+ *                           /             +- Select
+ *                          /              +- Literal
+ *                         /               +- This
+ *                        /                +- New
+ *                       /                 +- NamedArg
+ *                      /                  +- Apply
+ *                     /                   +- TypeApply
+ *                    /                    +- Super
+ *                   /                     +- Typed
+ *                  /                      +- Assign
+ *                 /                       +- Block
+ *  +- Parent ----+                        +- Lambda
+ *                 \                       +- If
+ *                  \                      +- Match
+ *                   \                     +- Try
+ *                    \                    +- Return
+ *                     \                   +- Repeated
+ *                      \                  +- Inlined
+ *                       \                 +- SelectOuter
+ *                        \                +- While
+ *                         \               +- DoWhile
+ *                          \
+ *                           \
+ *                         +- TypeTree ----+- Synthetic
+ *                         |               +- TypeIdent
+ *                         |               +- TermSelect
+ *                         |               +- TypeSelect
+ *                         |               +- Singleton
+ *  +- TypeOrBoundsTree ---+               +- Refined
+ *                         |               +- Applied
+ *                         |               +- Annotated
+ *                         |               +- And
+ *                         |               +- Or
+ *                         |               +- ByName
+ *                         |               +- TypeLambdaTree
+ *                         |               +- Bind
+ *                         |
+ *                         +- TypeBoundsTree
+ *                         +- SyntheticBounds
+ *
+ *  +- CaseDef
+ *
+ *  +- Pattern --+- Value
+ *               +- Bind
+ *               +- Unapply
+ *               +- Alternative
+ *               +- TypeTest
+ *
+ *
+ *                   +- NoPrefix
+ *  +- TypeOrBounds -+- TypeBounds
+ *                   |
+ *                   +- Type -------+- ConstantType
+ *                                  +- SymRef
+ *                                  +- TermRef
+ *                                  +- TypeRef
+ *                                  +- SuperType
+ *                                  +- Refinement
+ *                                  +- AppliedType
+ *                                  +- AnnotatedType
+ *                                  +- AndType
+ *                                  +- OrType
+ *                                  +- ByNameType
+ *                                  +- ParamRef
+ *                                  +- ThisType
+ *                                  +- RecursiveThis
+ *                                  +- RecursiveType
+ *                                  +- LambdaType[ParamInfo <: TypeOrBounds] -+- MethodType
+ *                                                                            +- PolyType
+ *                                                                            +- TypeLambda
+ *
+ *  +- ImportSelector -+- SimpleSelector
+ *                     +- RenameSelector
+ *                     +- OmitSelector
+ *
+ *  +- Id
+ *
+ *  +- Signature
+ *
+ *  +- Position
+ *
+ *  +- Constant
+ *
+ *  ```
+ */
+trait TastyCore {
+
+  type Context
+
+  // TODO: When bootstrapped, remove and use `Term | TypeTree` type directly
+  type Parent /* Term | TypeTree */
+
+  type Tree
+    type PackageClause <: Tree
+    type Statement <: Tree
+      type Import <: Statement
+      type Definition <: Statement
+        type ClassDef <: Definition
+        type TypeDef <: Definition
+        type DefDef <: Definition
+        type ValDef <: Definition
+        type PackageDef <: Definition
+      type Term <: Statement with Parent
+
+  type CaseDef
+
+  type Pattern
+
+  type TypeOrBoundsTree
+    type TypeTree <: TypeOrBoundsTree with Parent
+    type TypeBoundsTree <: TypeOrBoundsTree
+
+  type TypeOrBounds
+    type NoPrefix <: TypeOrBounds
+    type TypeBounds <: TypeOrBounds
+    type Type <: TypeOrBounds
+    type RecursiveType <: Type
+    // TODO can we add the bound back without an cake?
+    // TODO is LambdaType really needed? ParamRefExtractor could be split into more precise extractors
+    type LambdaType[ParamInfo /*<: TypeOrBounds*/] <: Type
+      type MethodType <: LambdaType[Type]
+      type PolyType <: LambdaType[TypeBounds]
+      type TypeLambda <: LambdaType[TypeBounds]
+
+
+  type ImportSelector
+
+  type Id
+
+  type Signature
+
+  type Position
+
+  type Constant
+
+}
